Rename modules (Form/Table -> StringifiedQuery/StructuredQuery) and refactor following it
