+++ src/ts/modules/StructuredQuery.ts
@@ -1,33 +1,35 @@
 import * as TypescriptFsa from 'typescript-fsa';
 import * as TypescriptFsaReducers from 'typescript-fsa-reducers';
 
+import {ObjectTable} from '../logic/table-data';
+
 export interface State {
-  text: string;
+  table: ObjectTable;
 }
 
 const initialReduceState: State = {
-  text: ''
+  table: []
 };
 
-function handler(state: State, payload: string): any {
-  return {...state, text: payload};
+function handler(_: State, table: ObjectTable): any {
+  return {table};
 }
 
 // ----------------------------------------------------------------------------
 // action types
 // ----------------------------------------------------------------------------
-const SET_TEXT = 'react-studies/Form/SET_TEXT';
+const SET_TABLE = 'react-studies/StructuredQuery/SET_TABLE';
 
 // ----------------------------------------------------------------------------
 // action creators
 // ----------------------------------------------------------------------------
 const actionCreator = TypescriptFsa.actionCreatorFactory();
 
-export const setFormText = actionCreator<string>(SET_TEXT);
+export const setTable = actionCreator<ObjectTable>(SET_TABLE);
 
 // ----------------------------------------------------------------------------
 // reducer
 // ----------------------------------------------------------------------------
 export default TypescriptFsaReducers.reducerWithInitialState(initialReduceState)
-    .case(setFormText, handler)
+    .case(setTable, handler)
     .build();
